+++ test/benchmarks/raphael/index.mustache
@@ -22,6 +22,7 @@
       <{{tag}} src="{{src}}" onload="preload(this)"></{{tag}}>
       {{/assets}}
     </div>
+    {{{pre}}}
     <div class="flex-row">
       <div class="px-1">
         <canvas id="canvas" width="600" height="600"></canvas>
