+++ test/benchmarks/raphael/simple/header.html
@@ -1,0 +1,17 @@
+<header>
+  Rendering a large number (<span id="numshapes"></span>) of simple shapes
+  <div>
+    <input id="optimize_caching" type="checkbox">
+    <label for="optimize_caching">Optimize caching</label>
+    <input id="object_stacking" type="checkbox">
+    <label for="object_stacking">Preserve object stacking</label>
+  </div>
+  <ul>
+    <li><a href="#" onclick="return reload(10)">10 shapes</a></li>
+    <li><a href="#" onclick="return reload(100)">100 shapes</a></li>
+    <li><a href="#" onclick="return reload(500)">500 shapes</a></li>
+    <li><a href="#" onclick="return reload(1000)">1000 shapes</a></li>
+    <li><a href="#" onclick="return reload(2000)">2000 shapes</a></li>
+    <li><a href="#" onclick="return reload(5000)">5000 shapes</a></li>
+  </ul>
+</header>
